feat: deprecate `litestar.contrib.prometheus`
