+++ docs/reference/plugins/prometheus.rst
@@ -1,0 +1,5 @@
+prometheus
+==========
+
+.. automodule:: litestar.plugins.prometheus
+    :members:
